port 8.1 changes

+++ modules/snmp/gateway/src/main/kotlin/com/mussonindustrial/embr/snmp/configuration/extensions/SnmpDeviceConfig.kt
@@ -33,7 +33,7 @@ data class SnmpConnectivityConfig(
     @DefaultValue("1000")
     @FormField(FormFieldType.NUMBER)
     @Required
-    val timeout: Long,
+    val timeout: Int,
 ) {
     fun validate(errors: ValidationErrors.Builder) =
         errors.apply {
